LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-fonts/math-script-level-and-math-style/math-script-level-001.tentative-expected.txt


FAIL Initial value of math-depth assert_equals: expected "0" but got ""
FAIL Inherited values of math-depth assert_equals: expected "11" but got ""
FAIL Specified math-depth: auto-add assert_equals: expected "10" but got ""
FAIL Specified math-depth: <integer> assert_equals: expected "11" but got ""
FAIL Specified math-depth: add(<integer>) assert_equals: expected "15" but got ""
FAIL Specified math-depth: invalid expressions assert_equals: expected "3" but got ""
FAIL Specified math-depth: calc() expressions assert_equals: expected "12" but got ""
